+++ dev_codes.py
@@ -1,17 +1,16 @@
-import json_read as jr
 import recipe_manager as rm
 
 ###############################################
 # dev code to iterate through all recipes
 # to check if any recipe fails to parse
-x = jr.recipe_loader()
+x = rm.recipe_loader()
 print(len(x))
 i = 0
 
 for iterator in x:
     try:
         y = rm.unique_items_required(iterator)
-        #print(i)
+        # print(i)
         i += 1
     except:
         print("failed", iterator)
