+++ recipe_manager.py
@@ -1,6 +1,6 @@
 import json
 import constants
-import json_read
+import os
 
 '''
     Method to identify unique ingredients required in an item
@@ -44,3 +44,42 @@ def unique_items_required(item_name):
 # unique_items_required("beacon")
 # unique_items_required("fire_charge")
 # print(type(x))
+
+'''
+    Loads all the recipe's file names
+    No parameter
+    returns list of filenames
+'''
+
+
+def recipe_loader():
+
+    # "w" creates new file in write mode overwrites if already exist
+    shapeless = list()
+    shaped = list()
+    f_recipe = list()
+    file_names = list()
+
+    # reads the files in the folder
+    for filename in os.listdir(constants.abs_recipe_dir):
+
+        abs_file_path = constants.abs_recipe_dir + filename
+        read_data = open(abs_file_path).read()  # File Handle
+        data = json.loads(read_data)  # Json object
+
+        # Categorising into shaped and shapeless recipes
+        if data["type"] == "crafting_shapeless":
+            shapeless.append(filename)
+
+        else:
+            shaped.append(filename)
+
+        # Listing all the recipes Available
+        if data["result"]:
+            fname = data["result"]["item"]
+            fname = fname.replace("minecraft:", "")
+            f_recipe.append(fname)
+        tmp_file_name = filename.replace(".json", "")
+        file_names.append(tmp_file_name)
+
+    return file_names
